int for numberOfPartiesHosted changed to long.

+++ src/main/java/building/StudentFlat.java
@@ -2,7 +2,7 @@ package building;
 public class StudentFlat extends Building {
     private long numberOfPartiesHosted;
 
-    public StudentFlat (String dateOfConstruction, int numberOfFloors, int numberOfRooms, int numberOfPartiesHosted){
+    public StudentFlat (String dateOfConstruction, int numberOfFloors, int numberOfRooms, long numberOfPartiesHosted){
         super(dateOfConstruction, numberOfFloors, numberOfRooms);
         this.numberOfPartiesHosted = numberOfPartiesHosted;
     }
